feat: Caching for guild endpoints

+++ neptunium-http/src/endpoints/guild/roles/update_role_hoist_positions.rs
@@ -11,7 +11,7 @@ use crate::{endpoints::Endpoint, request::Request};
 #[derive(Builder, Serialize, Copy, Clone, Debug)]
 pub struct UpdateGuildRoleHoistPositionsEntry {
     pub id: Id<RoleMarker>,
-    pub hoist_position: u64,
+    pub hoist_position: u16,
 }
 
 #[derive(Builder, Clone, Debug)]
